delete content view and test view
