+++ CoreDataCloudKitShare/MainView.swift
@@ -9,6 +9,26 @@
 import Foundation
 import SwiftUI
 
+class Theme {
+    static func navigationBarColors(background : UIColor?,
+       titleColor : UIColor? = nil, tintColor : UIColor? = nil ){
+        
+        let navigationAppearance = UINavigationBarAppearance()
+        navigationAppearance.configureWithOpaqueBackground()
+        navigationAppearance.backgroundColor = background ?? .clear
+        navigationAppearance.shadowColor = .clear
+        
+        navigationAppearance.titleTextAttributes = [.foregroundColor: titleColor ?? .black]
+        navigationAppearance.largeTitleTextAttributes = [.foregroundColor: titleColor ?? .black]
+       
+        UINavigationBar.appearance().standardAppearance = navigationAppearance
+        UINavigationBar.appearance().compactAppearance = navigationAppearance
+        UINavigationBar.appearance().scrollEdgeAppearance = navigationAppearance
+
+        UINavigationBar.appearance().tintColor = tintColor ?? titleColor ?? .black
+    }
+}
+
 struct MainView: View {
     @Environment(\.managedObjectContext) private var viewContext
     @FetchRequest(sortDescriptors: [SortDescriptor(\.namaPemilik)],
@@ -146,7 +166,7 @@ struct MainView: View {
             
             {
 
-                OnboardingViewPure(data: ContentView.sample, doneFunction: {
+                OnboardingViewPure(data: MainView.sample, doneFunction: {
                     
                     withAnimation{
                         UserDefaults.standard.set(true, forKey: "isStart")
